+++ apps-client/mobile/libs/react-query.ts
@@ -3,7 +3,7 @@ import { createReactQueryClient } from '@pple-today/api-client'
 import type { ApplicationApiSchema } from '@api/backoffice'
 import { environment } from '@app/env'
 
-import { getAuthSession } from './auth'
+import { getAuthSession } from './auth/session'
 
 const { fetchClient, queryClient } = createReactQueryClient<ApplicationApiSchema>(
   environment.EXPO_PUBLIC_BACKEND_BASE_URL
